dblhemi->dblvrs, trplhemi->trplvrs
After a conversation with Wayne, we replaced the confusing and misleading terminology "double hemistich" and "triple hemistich" (or "dblhemi" and "trplhemi", in the code) with "double verse" and "triple verse" (or "dblvrs" and "trplvrs" in the code). Isabella, will you please update the wiki documentation to reflect this change. Thanks!

John

Signed-off-by: John A. Walsh <jawalsh@indiana.edu>