Created separate module to handle Facsimile
Deleted Magnific pop-up JS jquery plugin.

Pushing the most crude version of what we have so far for Facs image viewer.
Currently is just placed as a FIXED element on right side of the page. User can zoom in and out.

Next step is to allow user to resize the fixed element