Put up documentation folder and added javascript documentation as markdown file.
Refactored alot of the modules. Writing the documentation really helped to organize things better. it was a mess before! Still a mess but not as bad.

Changed index.php file to work by removing the fancy php way of loading css/js. Now loading both types of files the old school way. Next is to change all .php files to do the same.